move to segments.rs
